migrate out of src, add components and types folders, rename auth-form component file, update imports
