+++ app/components/AuthForm.tsx
@@ -2,7 +2,7 @@
 import { Auth } from "@supabase/auth-ui-react";
 import { ThemeSupa } from "@supabase/auth-ui-shared";
 import { createClientComponentClient } from "@supabase/auth-helpers-nextjs";
-import { Database } from "../app/database.types";
+import { Database } from "../types/database.types";
 
 export default function AuthForm() {
   const supabase = createClientComponentClient<Database>();
