--- app/database.types.ts
@@ -1,49 +1,0 @@
-export type Json =
-  | string
-  | number
-  | boolean
-  | null
-  | { [key: string]: Json }
-  | Json[];
-
-export interface Database {
-  public: {
-    Tables: {
-      profiles: {
-        Row: {
-          id: string;
-          updated_at: string | null;
-          username: string | null;
-          full_name: string | null;
-          avatar_url: string | null;
-          website: string | null;
-        };
-        Insert: {
-          id: string;
-          updated_at?: string | null;
-          username?: string | null;
-          full_name?: string | null;
-          avatar_url?: string | null;
-          website?: string | null;
-        };
-        Update: {
-          id?: string;
-          updated_at?: string | null;
-          username?: string | null;
-          full_name?: string | null;
-          avatar_url?: string | null;
-          website?: string | null;
-        };
-      };
-    };
-    Views: {
-      [_ in never]: never;
-    };
-    Functions: {
-      [_ in never]: never;
-    };
-    Enums: {
-      [_ in never]: never;
-    };
-  };
-}
