--- src/app/page.tsx
@@ -1,7 +1,0 @@
-export default function Home() {
-  return (
-    <main className="flex min-h-screen flex-col items-center justify-between p-24">
-      PickleConnect
-    </main>
-  );
-}
